make rotate speed a float

+++ Assets/_Scripts/movable_object.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class movable_object : MonoBehaviour
 {
-    public int rotateSpeed = 3;
+    public float rotateSpeed = 3.0;
     // Start is called before the first frame update
     void Start()
     {
@@ -24,7 +24,7 @@ public class movable_object : MonoBehaviour
             transform.Rotate(0,0, rotateSpeed);
         } else if (Input.GetKey("e"))
         {
-            transform.Rotate(0, 0, (rotateSpeed * -1));
+            transform.Rotate(0, 0, (rotateSpeed * -1.0f));
 
         }
